patch overflow

+++ src/board/cell.rs
@@ -4,14 +4,15 @@ pub enum STATUS {
     ALIVE = 1
 }
 
+#[derive(Clone, Copy)]
 pub struct Cell {
-    pub x: usize,
-    pub y: usize,
+    pub x: i32,
+    pub y: i32,
     pub status: STATUS
 }
 
 impl Cell {
-    pub fn new(x: usize, y: usize, status: STATUS) -> Self {
+    pub fn new(x: i32, y: i32, status: STATUS) -> Self {
         Cell{x, y, status}
     }
 }
